pyproject.toml
[project]
name = "python-ai"
version = "0.1.0"
description = "AI Project"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai==1.12.1",
    "numpy>=2.0.2",
    "python-dotenv==1.1.0",
]
